+++ bitorch/__init__.py
@@ -5,7 +5,10 @@ from pathlib import Path
 from typing import List
 
 from .config import Config
+from .runtime_mode import RuntimeMode, runtime_mode_type, change_mode, pause_wrapping
+from .layers import convert
 
+mode: RuntimeMode = RuntimeMode.DEFAULT
 
 configs_by_name = {}
 
@@ -64,8 +67,8 @@ def add_config_args(parser: ArgumentParser) -> None:
     Args:
         parser (ArgumentParser): parser to add the arguments to
     """
-    for config in configs_by_name.values():
-        config.add_config_arguments(parser)
+    for config_ in configs_by_name.values():
+        config_.add_config_arguments(parser)
 
 
 def apply_args_to_configuration(args: Namespace) -> None:
@@ -74,5 +77,5 @@ def apply_args_to_configuration(args: Namespace) -> None:
     Args:
         args (Namespace): the cli configurations
     """
-    for config in configs_by_name.values():
-        config.apply_args_to_configuration(args)
+    for config_ in configs_by_name.values():
+        config_.apply_args_to_configuration(args)