+++ bitorch/layers/extensions/__init__.py
@@ -1,0 +1,9 @@
+"""
+This submodule contains objects needed to provide and manage custom layer implementations.
+"""
+
+from .layer_container import LayerContainer
+from .layer_implementation import DefaultImplementationMixin, CustomImplementationMixin
+from .layer_recipe import LayerRecipe
+from .layer_registration import LayerImplementation
+from .layer_registry import LayerRegistry