+++ docs/source/conf.py
@@ -44,6 +44,9 @@ extensions = [
     'nbsphinx_link',
 ]
 
+# Generate type hints
+autodoc_typehints = "description"
+
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
 
@@ -65,7 +68,7 @@ html_theme = 'alabaster'
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
+# html_static_path = ['_static']
 
 
 # # This is the expected signature of the handler for this event, cf doc